backend/llm_summary_service/requirements.txt
fastapi>=0.112.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
requests>=2.31.0,<3.0.0
openai>=1.35.0,<2.0.0
pytest>=8.2.0,<9.0.0
python-dotenv>=1.0.1,<2.0.0
